refactor: Integration library project structure
Part of YIO-Remote/remote-software#348
