+++ README.md
@@ -1,4 +1,4 @@
 # YIO Dock Repository
 
 For details about the YIO Dock Integration, please visit our documentation repository which can be found under
-https://github.com/YIO-Remote/documentation/wiki
+<https://github.com/YIO-Remote/documentation/wiki>
