+++ cpplint.sh
@@ -1,0 +1,1 @@
+cpplint --filter=-build/include_subdir --linelength=120 --root=src --recursive src
